save timestamp as float

+++ api/model/errormodel.go
@@ -6,6 +6,6 @@ type Error struct {
 	Title       string   `json:"title,omitempty" bson:"title"`
 	Type        string   `json:"type,omitempty" bson:"type"`
 	StackTrace  string   `json:"stacktrace,omitempty" bson:"stacktrace"`
-	Timestamp   int      `json:"timestamp,omitempty" bson:"timestamp"`
+	Timestamp   float64  `json:"timestamp,omitempty" bson:"timestamp"`
 	MetricsId   string   `json:"metricsid,omitempty" bson:"metricsId"`
 }
